+++ bundles/former/config/former.php
@@ -15,6 +15,13 @@
   // Whether Former should try to apply Validator rules as attributes
   'live_validation'   => true,
 
+  // Whether checkboxes should always be present in the POST data,
+  // no matter if you checked them or not
+  'push_checkboxes'   => false,
+
+  // The value a checkbox will have in the POST array if unchecked
+  'unchecked_value'   => '',
+
   // The class to be added to required fields
   'required_class'    => 'required',