+++ bundles/former/libraries/Fields/Select.php
@@ -51,11 +51,23 @@ class Select extends \Former\Field
   /**
    * Set the select options
    *
-   * @param  array $options  The options as an array
-   * @param  mixed $selected Facultative selected entry
+   * @param  array   $options      The options as an array
+   * @param  mixed   $selected     Facultative selected entry
+   * @param  boolean $valuesAsKeys Whether the array's values should be used as
+   *                               the option's values instead of the array's keys
    */
-  public function options($options, $selected = null)
+  public function options($_options, $selected = null, $valuesAsKeys = false)
   {
+    // Automatically fetch Lang objects for people who store translated options lists
+    if($_options instanceof \Laravel\Lang) {
+      $_options = $_options->get();
+    }
+
+    // If valuesAsKeys is true, use the values as keys
+    if($valuesAsKeys) {
+      foreach($_options as $v) $options[$v] = $v;
+    } else $options = $_options;
+
     $this->options = $options;
 
     if($selected) $this->value = $selected;
@@ -92,7 +104,17 @@ class Select extends \Former\Field
    */
   public function placeholder($placeholder)
   {
-    $this->placeholder = $placeholder;
+    $this->placeholder = Helpers::translate($placeholder);
+  }
+
+  /**
+   * Returns the current options in memory for manipulations
+   *
+   * @return array The current options array
+   */
+  public function getOptions()
+  {
+    return $this->options;
   }
 
   /**