+++ bundles/former/tests/ControlGroup.test.php
@@ -2,7 +2,7 @@
 use \Former\Former;
 
 // Stub class for Buttons
-class Buttons
+class Button
 {
   public static function normal($text)
   {
@@ -103,10 +103,26 @@ class ControlGroupTest extends FormerTests
     $this->assertEquals($matcher, $control);
   }
 
+  public function testEmptyInlineHelp()
+  {
+    $control = Former::text('foo')->inlineHelp(null)->__toString();
+    $matcher = $this->createMatcher();
+
+    $this->assertEquals($matcher, $control);
+  }
+
   public function testBlockHelp()
   {
     $control = Former::text('foo')->blockHelp('foo')->__toString();
     $matcher = $this->createMatcher(null, null, 'foo');
+
+    $this->assertEquals($matcher, $control);
+  }
+
+  public function testEmptyBlockHelp()
+  {
+    $control = Former::text('foo')->blockHelp(null)->__toString();
+    $matcher = $this->createMatcher();
 
     $this->assertEquals($matcher, $control);
   }
@@ -154,8 +170,8 @@ class ControlGroupTest extends FormerTests
   public function testPrependAppendMix()
   {
     $control = Former::text('foo')
-      ->prepend('@', Buttons::normal('foo'))
-      ->append('@', Buttons::normal('foo'))
+      ->prepend('@', Button::normal('foo'))
+      ->append('@', Button::normal('foo'))
       ->__toString();
     $matcher = $this->createPrependAppendMatcher(
       array('@', '<button type="button" class="btn">foo</button>'),
@@ -166,7 +182,7 @@ class ControlGroupTest extends FormerTests
 
   public function testPrependButton()
   {
-    $control1 = Former::text('foo')->prepend(Buttons::normal('Submit'))->__toString();
+    $control1 = Former::text('foo')->prepend(Button::normal('Submit'))->__toString();
     $control2 = Former::text('foo')->prepend('<button type="button" class="btn">Submit</button>')->__toString();
     $matcher = $this->createPrependAppendMatcher(array('<button type="button" class="btn">Submit</button>'));
 
@@ -180,8 +196,8 @@ class ControlGroupTest extends FormerTests
       ->state('error')
       ->inlineHelp('foo')
       ->blockHelp('bar')
-      ->prepend('@', '$', Buttons::normal('foo'))
-      ->append('@', '$', Buttons::normal('foo'))
+      ->prepend('@', '$', Button::normal('foo'))
+      ->append('@', '$', Button::normal('foo'))
       ->__toString();
     $matcher =
     '<div class="control-group error">'.