+++ bundles/former/CHANGELOG.md
@@ -1,7 +1,20 @@
 ## Changelog
 
-### 2.*.*
+### 2.4.0
 
+- **[add] Form openers are now objects too and accept chained methods**
+- [add] Add `unchecked_value` option to decide what value unchecked checkboxes have in the POST array
+- [add] Allow booleans to be passed to Checkable->check() on single items
+- [mod] Disable `push_checkbox` option by default
+- [fix] Automatically fetch Lang objects passed to `->options`
+
+### 2.3.0
+
+- **[add] Add `push_checkboxes` option which forces the submiting of unchecked fields to the POST array**
+
+### 2.2.0
+
+- **[add] Add `Former::file()` and `Former::files()` with methods `->max` and `->accept`**
 - [add] Add ability to set attributes for a label
 - [add] Add ability to set a placeholder option for select fields